fix:충전 모달
